requirements.txt
streamlit
google-generativeai
Pillow
ipython
